# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ApproveMessageRequest": {
        "additionalProperties": false,
        "description": "What the approver confirms (\u00a711.3 steps 1-2).",
        "properties": {
          "recipient_contact_point_id": {
            "format": "uuid",
            "title": "Recipient Contact Point Id",
            "type": "string"
          },
          "record_version": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Version"
          }
        },
        "required": [
          "recipient_contact_point_id"
        ],
        "title": "ApproveMessageRequest",
        "type": "object"
      },
      "ApproveMessageResponse": {
        "description": "What was approved, and what it produced.",
        "properties": {
          "approval_id": {
            "format": "uuid",
            "title": "Approval Id",
            "type": "string"
          },
          "message_revision_id": {
            "format": "uuid",
            "title": "Message Revision Id",
            "type": "string"
          },
          "recipient": {
            "title": "Recipient",
            "type": "string"
          },
          "record_version": {
            "format": "date-time",
            "title": "Record Version",
            "type": "string"
          },
          "revision_state": {
            "title": "Revision State",
            "type": "string"
          },
          "send_command_id": {
            "format": "uuid",
            "title": "Send Command Id",
            "type": "string"
          },
          "what_happens_next": {
            "title": "What Happens Next",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "message_revision_id",
          "send_command_id",
          "recipient",
          "revision_state",
          "what_happens_next",
          "record_version"
        ],
        "title": "ApproveMessageResponse",
        "type": "object"
      },
...
      "AttentionPage": {
        "description": "Approvals needing attention, oldest first.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttentionRow"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AttentionPage",
        "type": "object"
      },
      "AttentionRow": {
        "description": "One approval that no longer authorizes a send (\u00a77.5, \u00a78.4).",
        "properties": {
          "approval_expires_at": {
            "format": "date-time",
            "title": "Approval Expires At",
            "type": "string"
          },
          "approval_id": {
            "format": "uuid",
            "title": "Approval Id",
            "type": "string"
          },
          "approver_id": {
            "title": "Approver Id",
            "type": "string"
          },
          "message_revision_id": {
            "format": "uuid",
            "title": "Message Revision Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "recipient_contact_point_id": {
            "format": "uuid",
            "title": "Recipient Contact Point Id",
            "type": "string"
          },
          "record_version": {
            "format": "date-time",
            "title": "Record Version",
            "type": "string"
          },
          "trigger": {
            "$ref": "#/components/schemas/InvalidationTrigger"
          },
          "triggering_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triggering Id"
          }
        },
        "required": [
          "approval_id",
          "message_revision_id",
          "recipient_contact_point_id",
          "approver_id",
          "approval_expires_at",
          "trigger",
          "reason",
          "triggering_id",
          "record_version"
        ],
        "title": "AttentionRow",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvalidationTrigger": {
        "description": "Which \u00a78.4 condition stopped an approval authorizing a send.\n\n\u00a78.4 lists six changes that invalidate an approval \u2014 recipient, subject, body, material\npersonalization fact, product status, and claim version \u2014 plus the two lifecycle facts that\nalso stop one being usable: the approval's own state, and its expiry. The three pinned values\non the row cover all six of the \u00a78.4 six, which is why there are fewer members here than that\nlist has entries: the content hash covers recipient, subject, body, and personalization\ntogether, because any of them changing changes the hash.",
        "enum": [
          "not_approved",
          "expired",
          "revision_missing",
          "revision_retired",
          "content_changed",
          "recipient_changed",
          "product_status_superseded",
          "claim_set_superseded"
        ],
        "title": "InvalidationTrigger",
        "type": "string"
      },
...
      "RevokeApprovalRequest": {
        "additionalProperties": false,
        "description": "Why this approval is being revoked.",
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "record_version": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Version"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RevokeApprovalRequest",
        "type": "object"
      },
      "RevokeApprovalResponse": {
        "description": "What the revocation did.",
        "properties": {
          "approval_id": {
            "format": "uuid",
            "title": "Approval Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "record_version": {
            "format": "date-time",
            "title": "Record Version",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "state",
          "reason",
          "record_version"
        ],
        "title": "RevokeApprovalResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Application-owned sales workflow. Shadow mode; live outreach is gated.",
    "title": "Matrix Power Always-On AI Sales Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/review/approvals/{approval_id}/revoke": {
      "post": {
        "description": "Withdraw an approval (\u00a78.2's `approved -> revoked`, \u00a717.6).\n\n**`APPROVE_MESSAGE`, the same permission that granted it.** Withdrawing an authorization is\nthe same authority as giving one \u2014 a role that could revoke but not approve could stop any\noutreach it disliked, and one that could approve but not revoke could not undo its own\nmistake.\n\nA reason is required with no default: \u00a717.6 wants operational actions explicable, and a\nrevocation nobody can explain later is one a reviewer cannot distinguish from a fault.",
        "operationId": "revoke_approval_endpoint_api_review_approvals__approval_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Approval Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "mp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mp Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke an approval so it can never dispatch",
        "tags": [
          "review"
        ]
      }
    },
    "/api/review/attention/approvals": {
      "get": {
        "description": "\u00a77.5's flag. A read, so `requires` rather than `requires_mutation`.\n\nOnly approvals still in `approved` are scanned: one already revoked or rejected is not an\nattention item \u2014 somebody dealt with it \u2014 and listing them would bury the ones nobody has\nlooked at.",
        "operationId": "approvals_needing_attention_endpoint_api_review_attention_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "campaign_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campaign Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "mp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approvals that no longer authorize a send",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/review/revisions/{revision_id}/approve": {
      "post": {
        "description": "\u00a711.3 step 1, then the transaction.\n\nEvery refusal below happens before `approve_message` is called, so a rejected request writes\nnothing at all \u2014 there is no partial state for the caller to undo, and no audit event claiming\nan approval that did not happen.",
        "operationId": "approve_message_endpoint_api_review_revisions__revision_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "mp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mp Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve an exact message revision for an exact recipient",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}